requirements.txt
langchain>=0.1.16
langchain-community>=0.0.33
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit>=1.18.1
altair>=4
chromadb>=0.4.24
langchain-groq>=0.1.3
InstructorEmbedding>=1.0.1
sentence-transformers>=2.5.1